3-9-Django-Project/requirements.txt
Django==4.0.4
Pillow==9.1.1
boto3==1.23.10
Faker==13.12.0
psycopg2==2.9.3
selenium==4.2.0
gunicorn==20.1.0
djangorestframework==3.13.1
